Dialog for <, > operators
Validation works with negation again.
Added apprise v2 plugin to have simple dialog.
Dialog is used for operators greater and lower than to allow user set
own value.
In future both operators have to be controlled in validation and value
has to be added as new bin with ajax execute to server.